src/main/java/com/study/springbootaws/web/HelloController.java
package com.study.springbootaws.web;

import com.study.springbootaws.web.dto.HelloResponseDto;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RestController;

//JSON 을 반환하는 컨트롤러로 만들어줌
@RestController
public class HelloController {
    @GetMapping("/hello")   // = @RequestMapping(method = RequestMethod.GET)
    public String hello() {
        return "hello";
    }

    @GetMapping("/hello/dto")
    public HelloResponseDto helloDto(@RequestParam("name") String name, @RequestParam("amount") int amount) {
        return new HelloResponseDto(name, amount);
    }
}
